backend/users/serializers.py
from rest_framework import serializers
from .models import Follow, User

class RegisterSerializer(serializers.ModelSerializer):
    password = serializers.CharField(write_only=True, required=True)

    class Meta:
        model = User
        fields = ['id', 'username', 'email', 'password']

    def create(self, validated_data):
        user = User.objects.create_user(
            username=validated_data['username'],
            email=validated_data['email'],
            password=validated_data['password']
        )
        return user

class UserSerializer(serializers.ModelSerializer):
    followers_count = serializers.SerializerMethodField()
    following_count = serializers.SerializerMethodField()


    class Meta:
        model = User
        fields = ['id', 'username', 'email', 'bio', 'profile_picture', 'created_at', 'updated_at','followers_count', 'following_count']
        read_only_fields = ['id', 'created_at', 'updated_at']    

    def get_followers_count(self, obj):
        return obj.followers.count()
    
    def get_following_count(self,obj):
        return obj.following.count()


class FollowSerializer(serializers.ModelSerializer):
    class Meta:
        model = Follow
        fields = ['id', 'follower', 'following', 'created_at']
        read_only_fields = ['follower', 'created_at']        